+++ glade_forms/lib/src/model/model.dart
@@ -1,2 +1,4 @@
+export 'glade_composed_model.dart';
 export 'glade_metadata.dart';
 export 'glade_model.dart';
+export 'glade_model_base.dart';