+++ glade_forms/lib/src/widgets/widgets.dart
@@ -1,3 +1,4 @@
+export 'glade_composed_list_builder.dart';
 export 'glade_form_builder.dart';
 export 'glade_form_consumer.dart';
 export 'glade_form_debug_info.dart';